+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Fixed
 - compatibility with python3.5:
+  - fixed `TypeError` due to `start_engine` passing `pathlib.Path`
+    to `subprocess.Popen` for `cwd`
   - tests: fixed `AttributeError` due to unavailable `MagicMock.assert_called_once`
 
 ## [1.0.2] - 2020-05-05