+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - fixed `TypeError` due to `start_engine` passing `pathlib.Path`
     to `subprocess.Popen` for `cwd`
   - tests: fixed `AttributeError` due to unavailable `MagicMock.assert_called_once`
+  - tests: fixed `FileNotFoundError` due to strict `pathlib.Path.resolve`
 
 ## [1.0.2] - 2020-05-05
 ### Fixed