+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## [1.0.3] - 2020-06-18
 ### Fixed
 - compatibility with python3.5:
   - fixed `TypeError` due to `start_engine` passing `pathlib.Path`
@@ -63,7 +65,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.4.0] - 2017-10-31
 
-[Unreleased]: https://github.com/fphammerle/tooncher/compare/1.0.2...HEAD
+[Unreleased]: https://github.com/fphammerle/tooncher/compare/1.0.3...HEAD
+[1.0.3]: https://github.com/fphammerle/tooncher/compare/1.0.2...1.0.3
 [1.0.2]: https://github.com/fphammerle/tooncher/compare/1.0.1...1.0.2
 [1.0.1]: https://github.com/fphammerle/tooncher/compare/1.0.0...1.0.1
 [1.0.0]: https://github.com/fphammerle/tooncher/compare/0.4.1...1.0.0